refactor: move sqlite availability checks into history_db
This commit consolidates the sqlite3 availability checking and import logic into `plugin/framework/history_db.py`, deleting the single-purpose `plugin/framework/sqlite_available.py`. All imports have been updated accordingly across `legacy_ui.py` and `default_tools.py`.

The test file `test_sqlite_available.py` was also renamed to `test_history_db.py`, with its contents merged to run alongside the other history db tests. Build artifacts were also intentionally checked and cleaned up.
