--- plugin/framework/sqlite_available.py
@@ -1,24 +1,0 @@
-# WriterAgent - AI Writing Assistant for LibreOffice
-# Copyright (c) 2024 John Balis
-# Copyright (c) 2026 KeithCu (modifications and relicensing)
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-"""Global guard: sqlite3 availability. When False, chat history uses JSON fallback and web cache is disabled."""
-
-try:
-    import sqlite3
-    HAS_SQLITE = True
-except ImportError:
-    sqlite3 = None  # type: ignore[assignment]
-    HAS_SQLITE = False
